requirements.txt
requests>=2.25.0
pandas>=1.1.0
python-dateutil>=2.8.0
